moved Link to its own class
